backend/requirements.txt
python-dotenv==0.20.0
Flask==2.1.2
flask-smorest==0.38.0
flask-sqlalchemy==2.5.1
flask_marshmallow==0.14.0
marshmallow==3.16.0
marshmallow-sqlalchemy==0.28.0
python-magic==0.4.27
psycopg2==2.9.3
Flask-Migrate==3.1.0
marshmallow_enum==1.5.1